Dissertation/requirements.txt
numpy
pandas
scikit-learn
matplotlib
nltk
spacy
lightgbm
tensorflow==2.17.0
torch
seaborn
wordcloud
emoji==1.4.1
openpyxl
tf-keras